Refactor streaming to util so toolbox.base doesn't depend on datatype code.
